requirements.txt
biopython
pandas
freesasa
numpy
